extract context from analyze command
